Reorder components to add a properly usable ResizeHandle for mobile
As the menu should be resizable through dragging the whole container
instead of just a flimsy handle, it was needed to place the handle
behind (or, around) the displayed elements.

On another note: This also eases the usage for multiple menus (main and
secondary) by reducing the need to write duplicate code for the body and
header of the offcanvas.